Working Website Class - with Testing
Revamped the Website.Java class to with working methods for processing a websites mobile friendly test and seo eval criteria. Tested the Website object is parcelable and can be passed to another activity via intent extra.

Also changed user agent on Rankings to Bot, to inform Google that the request is sent from a bot.

Still need to update the other activities to use the website class.